feat: full challenge solution using tf modules
