+++ terraform/variables.tf
@@ -1,0 +1,27 @@
+variable "REGION" {
+  type    = string
+  default = "eu-west-1"
+}
+
+variable "PROFILE_NAME" {
+  type    = string
+  default = ""
+}
+
+# full file path
+variable "CREDENTIALS_FILE" {
+  type    = string
+  default = ""
+}
+
+## ---------------------------
+
+variable "MY_TAG" {
+  type    = string
+  default = "gabtec"
+}
+
+variable "ROUTE53_ZONE_ID" {
+  type    = string
+  default = ""
+}
